+++ consensus/qbft/src/main/java/org/hyperledger/besu/consensus/qbft/network/QbftGossip.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.consensus.qbft.core.network;
+package org.hyperledger.besu.consensus.qbft.network;
 
 import org.hyperledger.besu.consensus.common.bft.Gossiper;
 import org.hyperledger.besu.consensus.common.bft.network.ValidatorMulticaster;
